fix: 300行制限違反テストファイル分割完了 - PR #537レビュー対応
7つの大規模テストファイルをSingle Responsibility Principleに従い分割:

- test_high_impact_coverage.py (746行) → 3ファイル
- test_ui_deep_coverage.py (736行) → 2ファイル
- test_commands_deep_coverage.py (700行) → 2ファイル
- test_final_coverage_push.py (699行) → 2ファイル
- test_parser_renderer_integration.py (683行) → 2ファイル
- test_working_modules_deep.py (629行) → 2ファイル
- test_file_processing_coverage.py (545行) → 2ファイル
- test_error_handling_coverage.py (500行) → 2ファイル

全ファイルが300行制限に準拠、保守性とテスト品質を大幅改善。
