+++ tests/test_convert_check_processor.py
@@ -1,4 +1,4 @@
-"""Commands Deep Coverage Tests
+"""Convert Command and Check Syntax Deep Coverage Tests
 
 深度カバレッジ向上のためのCommands系モジュールテスト。
 target: 0%カバレッジのcommands系モジュールを大幅改善。
@@ -520,181 +520,3 @@ Regular paragraph text.
             except Exception:
                 # その他のエラーも許容（依存関係問題など）
                 pass
-
-
-class TestConvertValidatorDeep:
-    """ConvertValidator深度テスト - 50行 0%→70%+目標"""
-
-    def test_convert_validator_initialization(self):
-        """ConvertValidator初期化テスト"""
-        from kumihan_formatter.commands.convert.convert_validator import (
-            ConvertValidator,
-        )
-
-        validator = ConvertValidator()
-        assert validator is not None
-
-    def test_convert_validator_valid_input(self):
-        """有効入力の検証テスト"""
-        from kumihan_formatter.commands.convert.convert_validator import (
-            ConvertValidator,
-        )
-
-        validator = ConvertValidator()
-
-        # 有効なパラメータでの検証
-        try:
-            if hasattr(validator, "validate"):
-                result = validator.validate("input.txt", "output/")
-                assert isinstance(result, (bool, dict, tuple))
-            elif hasattr(validator, "validate_input"):
-                result = validator.validate_input("input.txt")
-                assert isinstance(result, (bool, dict))
-
-        except Exception:
-            # 依存関係の問題は許容
-            pass
-
-    def test_convert_validator_invalid_input(self):
-        """無効入力の検証テスト"""
-        from kumihan_formatter.commands.convert.convert_validator import (
-            ConvertValidator,
-        )
-
-        validator = ConvertValidator()
-
-        # 無効なパラメータでの検証
-        invalid_cases = [
-            ("", ""),  # 空文字
-            (None, None),  # None
-            ("nonexistent.txt", "invalid/path/"),  # 存在しないパス
-        ]
-
-        for input_file, output_path in invalid_cases:
-            try:
-                if hasattr(validator, "validate"):
-                    result = validator.validate(input_file, output_path)
-                    # エラー検出または False を期待
-                    if isinstance(result, bool):
-                        assert not result or result  # どちらでも許容
-
-            except (ValueError, TypeError, FileNotFoundError):
-                # 期待される例外
-                pass
-            except Exception:
-                # 依存関係エラーは許容
-                pass
-
-
-class TestConvertWatcherDeep:
-    """ConvertWatcher深度テスト - 77行 0%→60%+目標"""
-
-    def test_convert_watcher_initialization(self):
-        """ConvertWatcher初期化テスト"""
-        from kumihan_formatter.commands.convert.convert_watcher import ConvertWatcher
-
-        # モックの依存関係で初期化
-        mock_processor = Mock()
-        mock_validator = Mock()
-
-        try:
-            watcher = ConvertWatcher(mock_processor, mock_validator)
-            assert watcher is not None
-
-            # 基本属性確認
-            assert hasattr(watcher, "start") or hasattr(watcher, "watch")
-
-        except Exception:
-            # 依存関係の問題は許容
-            pass
-
-    def test_convert_watcher_start_stop(self):
-        """ウォッチャー開始・停止テスト"""
-        from kumihan_formatter.commands.convert.convert_watcher import ConvertWatcher
-
-        mock_processor = Mock()
-        mock_validator = Mock()
-
-        try:
-            watcher = ConvertWatcher(mock_processor, mock_validator)
-
-            # 開始テスト
-            if hasattr(watcher, "start"):
-                # 短時間でテスト
-                with patch("time.sleep"):  # sleep をモック
-                    try:
-                        watcher.start("test.txt", "output/")
-                    except KeyboardInterrupt:
-                        pass  # 停止テスト
-                    except Exception:
-                        pass  # 依存関係エラーは許容
-
-            # 停止テスト
-            if hasattr(watcher, "stop"):
-                watcher.stop()
-
-        except Exception:
-            # 依存関係の問題は許容
-            pass
-
-
-class TestSampleCommandDeep:
-    """SampleCommand深度テスト - 小さいが重要"""
-
-    def test_sample_command_basic(self):
-        """SampleCommand基本テスト"""
-        from kumihan_formatter.commands.sample_command import SampleCommand
-
-        try:
-            sample = SampleCommand()
-            assert sample is not None
-
-            # サンプル生成テスト
-            if hasattr(sample, "execute"):
-                result = sample.execute()
-                assert isinstance(result, (str, Path))
-            elif hasattr(sample, "generate"):
-                result = sample.generate()
-                assert isinstance(result, (str, Path))
-            elif hasattr(sample, "create_sample"):
-                result = sample.create_sample()
-                assert isinstance(result, (str, Path))
-
-        except ImportError:
-            pytest.skip("SampleCommand module not available")
-        except Exception:
-            # 依存関係エラーは許容
-            pass
-
-    def test_sample_command_output(self):
-        """サンプル出力テスト"""
-        from kumihan_formatter.commands.sample_command import SampleCommand
-
-        try:
-            sample = SampleCommand()
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                output_file = f.name
-
-            try:
-                # ファイル出力テスト
-                if hasattr(sample, "execute"):
-                    result = sample.execute(output_file.replace(".txt", ""))
-                    assert isinstance(result, Path)
-                elif hasattr(sample, "save_to"):
-                    sample.save_to(output_file)
-                    assert Path(output_file).exists()
-                elif hasattr(sample, "write_sample"):
-                    sample.write_sample(output_file)
-                    assert Path(output_file).exists()
-
-            except Exception:
-                # 依存関係エラーは許容
-                pass
-            finally:
-                Path(output_file).unlink(missing_ok=True)
-
-        except ImportError:
-            pytest.skip("SampleCommand module not available")
