+++ tests/test_error_stream_config.py
@@ -1,4 +1,4 @@
-"""File Processing Coverage Tests
+"""Error Handling, Stream Processing and Config File Tests
 
 Focus on file I/O, encoding handling, and related functionality.
 Target: Increase file processing module coverage significantly.
@@ -9,236 +9,6 @@ import tempfile
 from pathlib import Path
 
 import pytest
-
-
-class TestFileUtilities:
-    """Test file utility functions"""
-
-    def test_file_reader_basic(self):
-        """Test basic file reading functionality"""
-        try:
-            from kumihan_formatter.core.utilities.file_reader import FileReader
-
-            reader = FileReader()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        # Create temporary file
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
-            f.write("Test content\nSecond line\n")
-            temp_path = f.name
-
-        try:
-            # Test file reading
-            content = reader.read_file(temp_path)
-            assert isinstance(content, str)
-            assert "Test content" in content
-            assert "Second line" in content
-
-            # Test encoding detection
-            try:
-                encoding = reader.detect_encoding(temp_path)
-                assert isinstance(encoding, str)
-            except AttributeError as ae:
-                # Method might not exist
-                pass
-
-        except ImportError as ie:
-            # Module might not exist
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-        finally:
-            Path(temp_path).unlink(missing_ok=True)
-
-    def test_file_writer_basic(self):
-        """Test basic file writing functionality"""
-        try:
-            from kumihan_formatter.core.utilities.file_writer import FileWriter
-
-            writer = FileWriter()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Create temporary file path
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                temp_path = f.name
-
-            # Test file writing
-            test_content = "Test output content\nMultiple lines\n"
-            writer.write_file(temp_path, test_content)
-
-            # Verify content was written
-            with open(temp_path, "r", encoding="utf-8") as f:
-                written_content = f.read()
-                assert written_content == test_content
-
-            # Test encoding specification
-            try:
-                writer.write_file(temp_path, test_content, encoding="utf-8")
-            except TypeError:
-                # Encoding parameter might not be supported
-                pass
-
-            Path(temp_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            # Module might not exist
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_encoding_utilities(self):
-        """Test encoding detection and handling"""
-        try:
-            from kumihan_formatter.core.utilities.encoding_utils import EncodingDetector
-
-            detector = EncodingDetector()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Test various encodings
-            test_cases = [
-                ("utf-8", "UTF-8 text content"),
-                ("utf-8", "日本語テキスト"),  # Japanese text
-                ("ascii", "ASCII text"),
-            ]
-
-            for encoding, content in test_cases:
-                # Create temporary file with specific encoding
-                with tempfile.NamedTemporaryFile(
-                    mode="w", encoding=encoding, suffix=".txt", delete=False
-                ) as f:
-                    f.write(content)
-                    temp_path = f.name
-
-                try:
-                    # Test encoding detection
-                    detected = detector.detect_encoding(temp_path)
-                    assert isinstance(detected, str)
-
-                    # Test reading with detected encoding
-                    read_content = detector.read_with_encoding(temp_path, detected)
-                    assert content in read_content
-
-                except AttributeError as ae:
-                    # Methods might not exist
-                    pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-                finally:
-                    Path(temp_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            # Module might not exist
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-
-class TestFileConverter:
-    """Test file conversion functionality"""
-
-    def test_file_converter_basic(self):
-        """Test basic file conversion"""
-        try:
-            from kumihan_formatter.core.file_converter import FileConverter
-
-            converter = FileConverter()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Create input file
-            input_content = """# Test Document
-
-This is a test document with:
-- List items
-- Multiple paragraphs
-
-;;;highlight;;; Important section ;;;
-"""
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                f.write(input_content)
-                input_path = f.name
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".html", delete=False
-            ) as f:
-                output_path = f.name
-
-            try:
-                # Test conversion
-                converter.convert_file(input_path, output_path)
-
-                # Verify output exists
-                assert Path(output_path).exists()
-
-                # Verify output content
-                with open(output_path, "r", encoding="utf-8") as f:
-                    output_content = f.read()
-                    assert isinstance(output_content, str)
-                    assert len(output_content) > 0
-
-            except AttributeError as ae:
-                # Methods might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-            finally:
-                Path(input_path).unlink(missing_ok=True)
-                Path(output_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            # Module might not exist
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_batch_file_converter(self):
-        """Test batch file conversion"""
-        try:
-            from kumihan_formatter.core.batch_converter import BatchConverter
-
-            converter = BatchConverter()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Create multiple input files
-            test_files = []
-            for i in range(3):
-                content = f"# Document {i}\n\nContent for document {i}."
-                with tempfile.NamedTemporaryFile(
-                    mode="w", suffix=".txt", delete=False
-                ) as f:
-                    f.write(content)
-                    test_files.append(f.name)
-
-            try:
-                # Test batch conversion
-                output_dir = tempfile.mkdtemp()
-                converter.convert_batch(test_files, output_dir)
-
-                # Verify outputs
-                output_files = list(Path(output_dir).glob("*.html"))
-                assert len(output_files) > 0
-
-            except AttributeError as ae:
-                # Methods might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-            finally:
-                # Cleanup
-                for file_path in test_files:
-                    Path(file_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            # Module might not exist
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
 
 
 class TestErrorHandling:
