--- tests/test_high_impact_coverage.py
@@ -1,746 +1,0 @@
-"""High Impact Coverage Tests
-
-Focused tests on modules with highest coverage potential.
-Targets specific methods and code paths for maximum coverage gain.
-"""
-
-import tempfile
-from pathlib import Path
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-
-class TestMainParserHighImpact:
-    """High impact tests for main parser functionality"""
-
-    def test_parser_parse_method_comprehensive(self):
-        """Test parser.parse() method comprehensively"""
-        from kumihan_formatter.parser import Parser
-
-        parser = Parser()
-
-        # Test real parsing scenarios that exercise code paths
-        scenarios = [
-            # Basic text
-            "Hello world",
-            # Headings
-            "# Heading 1\nContent",
-            "## Heading 2\nMore content",
-            # Kumihan syntax
-            ";;;highlight;;; Important text ;;;",
-            ";;;box;;; Boxed content ;;;",
-            # Multi-line content
-            """Line 1
-Line 2
-Line 3""",
-            # Mixed content
-            """# Title
-
-Some paragraph text.
-
-;;;highlight;;; Special content ;;;
-
-Another paragraph.""",
-        ]
-
-        for scenario in scenarios:
-            try:
-                # This should exercise the main parse method
-                result = parser.parse(scenario)
-
-                # Verify result structure
-                assert result is not None
-                if hasattr(result, "__iter__"):
-                    # Should be iterable (list of nodes)
-                    node_count = len(list(result))
-                    assert node_count >= 0
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                # Some scenarios may fail due to incomplete implementation
-                pytest.skip(
-                    f"Parse scenario not available: {scenario[:20]}... - {str(e)[:50]}"
-                )
-
-    def test_parser_internal_methods(self):
-        """Test parser internal methods for coverage"""
-        from kumihan_formatter.parser import Parser
-
-        parser = Parser()
-
-        # Test line-by-line parsing if available
-        test_lines = [
-            "Simple line",
-            "# Heading line",
-            ";;;syntax;;; line ;;;",
-            "",  # Empty line
-            "  Indented line",
-        ]
-
-        for line in test_lines:
-            try:
-                # Try various internal methods that might exist
-                if hasattr(parser, "parse_line"):
-                    result = parser.parse_line(line)
-                    assert result is not None
-
-                if hasattr(parser, "process_line"):
-                    result = parser.process_line(line)
-                    assert result is not None
-
-                if hasattr(parser, "classify_line"):
-                    result = parser.classify_line(line)
-                    assert result is not None
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestMainRendererHighImpact:
-    """High impact tests for main renderer functionality"""
-
-    def test_renderer_render_method_comprehensive(self):
-        """Test renderer.render() method comprehensively"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Create diverse node scenarios
-        node_scenarios = [
-            # Single paragraph
-            [Node("p", "Simple paragraph")],
-            # Heading
-            [Node("h1", "Main Title")],
-            # Multiple nodes
-            [
-                Node("h1", "Title"),
-                Node("p", "Paragraph 1"),
-                Node("p", "Paragraph 2"),
-            ],
-            # Nested structure
-            [Node("div", [Node("h2", "Section"), Node("p", "Content")])],
-            # Complex structure
-            [
-                Node(
-                    "div",
-                    [
-                        Node("h1", "Document Title"),
-                        Node(
-                            "div",
-                            [
-                                Node("p", "Introduction"),
-                                Node(
-                                    "ul",
-                                    [
-                                        Node("li", "Item 1"),
-                                        Node("li", "Item 2"),
-                                    ],
-                                ),
-                            ],
-                        ),
-                    ],
-                )
-            ],
-        ]
-
-        for nodes in node_scenarios:
-            try:
-                # This should exercise the main render method
-                result = renderer.render(nodes)
-
-                # Verify result
-                assert isinstance(result, str)
-                assert len(result) > 0
-
-                # Basic HTML structure checks
-                if nodes[0].type in ["p", "h1", "h2", "div"]:
-                    # Should contain some representation of the node type
-                    assert (
-                        nodes[0].type in result.lower()
-                        or f"<{nodes[0].type}>" in result
-                        or nodes[0].content in result
-                    )
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(
-                    f"Render scenario not available: {[n.type for n in nodes]} - {str(e)[:50]}"
-                )
-
-    def test_renderer_template_system(self):
-        """Test renderer template system"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-        simple_nodes = [Node("p", "Test content")]
-
-        # Test different templates if available
-        templates = ["default", "minimal", "detailed", "compact"]
-
-        for template in templates:
-            try:
-                if hasattr(renderer, "set_template"):
-                    renderer.set_template(template)
-                    result = renderer.render(simple_nodes)
-                    assert isinstance(result, str)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test template options
-        try:
-            if hasattr(renderer, "set_options"):
-                renderer.set_options(
-                    {"include_css": True, "minify": False, "encoding": "utf-8"}
-                )
-                result = renderer.render(simple_nodes)
-                assert isinstance(result, str)
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestConfigSystemHighImpact:
-    """High impact tests for configuration system"""
-
-    def test_config_manager_comprehensive_usage(self):
-        """Test comprehensive config manager usage"""
-        from kumihan_formatter.config.config_manager import ConfigManager
-
-        config_manager = ConfigManager()
-
-        # Test various configuration scenarios
-        config_scenarios = [
-            # Basic settings
-            {"output_format": "html", "encoding": "utf-8"},
-            # Template settings
-            {"template": "default", "include_css": True},
-            # Parser settings
-            {"strict_mode": False, "allow_html": True},
-            # Renderer settings
-            {"minify_output": False, "pretty_print": True},
-            # File settings
-            {"backup_files": True, "output_dir": "/tmp/test"},
-        ]
-
-        for config in config_scenarios:
-            try:
-                # Test loading configuration
-                config_manager.load_config(config)
-
-                # Test getting values
-                for key, expected_value in config.items():
-                    actual_value = config_manager.get(key)
-                    # Value should be retrievable (may be transformed)
-                    assert actual_value is not None or expected_value is None
-
-                # Test validation
-                if hasattr(config_manager, "validate"):
-                    is_valid = config_manager.validate()
-                    assert isinstance(is_valid, bool)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test config merging
-        try:
-            base_config = {"theme": "default", "debug": False}
-            overlay_config = {"debug": True, "verbose": True}
-
-            config_manager.load_config(base_config)
-            config_manager.merge_config(overlay_config)
-
-            # Debug should be overridden
-            assert config_manager.get("debug") == True
-            # Theme should remain
-            assert config_manager.get("theme") == "default"
-            # New value should be added
-            assert config_manager.get("verbose") == True
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-    def test_extended_config_functionality(self):
-        """Test extended config functionality"""
-        from kumihan_formatter.config.extended_config import ExtendedConfig
-
-        config = ExtendedConfig()
-
-        # Test advanced configuration features
-        try:
-            # Environment variable integration
-            config.load_from_env("KUMIHAN_")
-
-            # Configuration validation
-            config.set("output_format", "html")
-            config.set("template", "default")
-
-            # Validate settings
-            validation_result = config.validate_all()
-            assert isinstance(validation_result, (bool, dict, list))
-
-            # Export configuration
-            exported = config.export()
-            assert isinstance(exported, dict)
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-        # Test configuration sections
-        sections = ["parser", "renderer", "output", "files"]
-        for section in sections:
-            try:
-                config.create_section(section)
-                config.set_section_value(section, "enabled", True)
-
-                section_config = config.get_section(section)
-                assert isinstance(section_config, dict)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestFileOperationsHighImpact:
-    """High impact tests for file operations"""
-
-    def test_file_operations_core_comprehensive(self):
-        """Test file operations core comprehensive functionality"""
-        from kumihan_formatter.core.file_operations_core import FileOperationsCore
-
-        file_ops = FileOperationsCore()
-
-        # Create test files for operations
-        test_files = []
-        try:
-            # Create multiple test files
-            for i in range(3):
-                with tempfile.NamedTemporaryFile(
-                    mode="w", suffix=f"_test_{i}.txt", delete=False
-                ) as tmp:
-                    tmp.write(f"Test content {i}\nLine 2\nLine 3")
-                    test_files.append(tmp.name)
-
-            # Test batch operations
-            for file_path in test_files:
-                try:
-                    # Test reading
-                    content = file_ops.read_file(file_path)
-                    assert isinstance(content, str)
-                    assert f"Test content" in content
-
-                    # Test file info
-                    info = file_ops.get_file_info(file_path)
-                    assert isinstance(info, dict)
-
-                    # Test file validation
-                    is_valid = file_ops.validate_file(file_path)
-                    assert isinstance(is_valid, bool)
-
-                except (
-                    AttributeError,
-                    NotImplementedError,
-                    TypeError,
-                    ValueError,
-                    FileNotFoundError,
-                ) as e:
-                    pytest.skip(f"Method or operation not available: {e}")
-
-            # Test batch processing
-            try:
-                batch_result = file_ops.process_files(test_files)
-                assert batch_result is not None
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        finally:
-            # Cleanup
-            for file_path in test_files:
-                Path(file_path).unlink(missing_ok=True)
-
-    def test_file_path_utilities_comprehensive(self):
-        """Test file path utilities comprehensive functionality"""
-        from kumihan_formatter.core.file_path_utilities import FilePathUtilities
-
-        utils = FilePathUtilities()
-
-        # Test path operations
-        test_paths = [
-            "/home/user/document.txt",
-            "relative/path/file.md",
-            "C:\\Windows\\Path\\file.txt",
-            "./current/dir/file.kumihan",
-            "../parent/dir/file.html",
-        ]
-
-        for path in test_paths:
-            try:
-                # Test path normalization
-                normalized = utils.normalize_path(path)
-                assert isinstance(normalized, str)
-
-                # Test path validation
-                is_valid = utils.validate_path(path)
-                assert isinstance(is_valid, bool)
-
-                # Test extension handling
-                extension = utils.get_extension(path)
-                assert isinstance(extension, str)
-
-                # Test directory extraction
-                directory = utils.get_directory(path)
-                assert isinstance(directory, str)
-
-                # Test filename extraction
-                filename = utils.get_filename(path)
-                assert isinstance(filename, str)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test path construction
-        try:
-            constructed = utils.join_paths("/base/path", "subdir", "file.txt")
-            assert isinstance(constructed, str)
-            assert "file.txt" in constructed
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestKeywordParsingHighImpact:
-    """High impact tests for keyword parsing"""
-
-    def test_keyword_parser_comprehensive_parsing(self):
-        """Test keyword parser comprehensive parsing"""
-        from kumihan_formatter.core.keyword_parser import KeywordParser
-
-        parser = KeywordParser()
-
-        # Test comprehensive keyword scenarios
-        keyword_scenarios = [
-            # Basic highlight
-            ";;;highlight;;; Important text ;;;",
-            # Box syntax
-            ";;;box;;; Boxed content ;;;",
-            # Footnote syntax
-            ";;;footnote;;; Footnote text ;;;",
-            # Multiple keywords in text
-            "Start ;;;highlight;;; middle ;;; end",
-            # Nested content
-            ";;;box;;; Content with ;;;highlight;;; nested ;;; syntax ;;;",
-            # Ruby notation
-            "｜漢字《かんじ》",
-            # Footnote parentheses
-            "Text with ((footnote content)) embedded",
-            # Complex mixed content
-            """Document with ;;;highlight;;; important ;;; information.
-
-And ｜漢字《かんじ》 ruby text.
-
-Plus ((a footnote)) reference.""",
-        ]
-
-        for scenario in keyword_scenarios:
-            try:
-                # Parse the scenario
-                result = parser.parse(scenario)
-                assert result is not None
-
-                # Should return structured data
-                if hasattr(result, "__len__"):
-                    assert len(result) >= 0
-
-                # Test keyword detection
-                detected_keywords = parser.find_keywords(scenario)
-                assert isinstance(detected_keywords, (list, tuple, set))
-
-                # Test keyword extraction
-                extracted = parser.extract_keywords(scenario)
-                assert extracted is not None
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-    def test_marker_parser_comprehensive_detection(self):
-        """Test marker parser comprehensive detection"""
-        from kumihan_formatter.core.keyword_parsing.marker_parser import MarkerParser
-
-        parser = MarkerParser()
-
-        # Test various marker patterns
-        marker_tests = [
-            (";;;highlight;;;", "highlight"),
-            (";;;box;;;", "box"),
-            (";;;footnote;;;", "footnote"),
-            ("((", "footnote_start"),
-            ("))", "footnote_end"),
-            ("｜", "ruby_start"),
-            ("《", "reading_start"),
-            ("》", "reading_end"),
-        ]
-
-        for marker, expected_type in marker_tests:
-            try:
-                # Test marker detection
-                detected = parser.detect_marker(marker)
-                assert detected is not None
-
-                # Test marker classification
-                marker_type = parser.classify_marker(marker)
-                assert marker_type is not None
-
-                # Test marker validation
-                is_valid = parser.validate_marker(marker)
-                assert isinstance(is_valid, bool)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test full text parsing for markers
-        full_text = "Text ;;;highlight;;; content ;;; and ｜ruby《reading》 notation."
-        try:
-            all_markers = parser.find_all_markers(full_text)
-            assert isinstance(all_markers, (list, tuple))
-
-            # Should find multiple markers
-            if len(all_markers) > 0:
-                for marker_info in all_markers:
-                    # Each marker should have position and type info
-                    assert isinstance(marker_info, (dict, tuple, list))
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestLoggerHighImpact:
-    """High impact tests for logging system"""
-
-    def test_logger_comprehensive_functionality(self):
-        """Test logger comprehensive functionality"""
-        from kumihan_formatter.core.utilities.logger import get_logger
-
-        # Test multiple logger instances
-        logger_names = ["parser", "renderer", "config", "file_ops", "main"]
-
-        for logger_name in logger_names:
-            try:
-                logger = get_logger(logger_name)
-                assert logger is not None
-
-                # Test all logging levels
-                log_levels = ["debug", "info", "warning", "error", "critical"]
-                for level in log_levels:
-                    if hasattr(logger, level):
-                        log_method = getattr(logger, level)
-
-                        # Test basic logging
-                        log_method(f"Test {level} message")
-
-                        # Test logging with extra data
-                        log_method(f"Test {level} with data", extra={"key": "value"})
-
-                        # Test exception logging
-                        if level in ["error", "critical"]:
-                            try:
-                                raise ValueError("Test exception")
-                            except ValueError as e:
-                                log_method(f"Exception occurred: {e}", exc_info=True)
-
-                # Test logger configuration
-                if hasattr(logger, "setLevel"):
-                    logger.setLevel("INFO")
-
-                if hasattr(logger, "addHandler"):
-                    # Create a test handler
-                    import logging
-
-                    handler = logging.StreamHandler()
-                    logger.addHandler(handler)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-    def test_structured_logger_base_comprehensive(self):
-        """Test structured logger base comprehensive functionality"""
-        import logging
-
-        from kumihan_formatter.core.utilities.structured_logger_base import (
-            StructuredLoggerBase,
-        )
-
-        # Create base logger
-        base_logger = logging.getLogger("test_structured_comprehensive")
-        structured_logger = StructuredLoggerBase(base_logger)
-
-        # Test structured logging scenarios
-        logging_scenarios = [
-            # Performance logging
-            {
-                "method": "log_performance",
-                "args": ("parse_operation", 0.125),
-                "kwargs": {"lines_processed": 150, "nodes_created": 45},
-            },
-            # File operation logging
-            {
-                "method": "log_file_operation",
-                "args": ("read", "/path/to/file.txt", True),
-                "kwargs": {"size_bytes": 2048, "encoding": "utf-8"},
-            },
-            # Error logging with context
-            {
-                "method": "log_error_with_context",
-                "args": ("Parse error occurred",),
-                "kwargs": {
-                    "line_number": 42,
-                    "file_path": "/test/file.txt",
-                    "error_type": "SyntaxError",
-                },
-            },
-            # State change logging
-            {
-                "method": "log_state_change",
-                "args": ("parser_mode", "strict", "permissive"),
-                "kwargs": {"reason": "user_configuration"},
-            },
-        ]
-
-        for scenario in logging_scenarios:
-            try:
-                method_name = scenario["method"]
-                if hasattr(structured_logger, method_name):
-                    method = getattr(structured_logger, method_name)
-
-                    args = scenario.get("args", ())
-                    kwargs = scenario.get("kwargs", {})
-
-                    # Execute the logging method
-                    method(*args, **kwargs)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test context filtering
-        try:
-            # Test with sensitive data that should be filtered
-            sensitive_context = {
-                "username": "testuser",
-                "password": "secret123",  # Should be filtered
-                "api_key": "abc123xyz",  # Should be filtered
-                "file_path": "/safe/path",
-                "token": "bearer_token",  # Should be filtered
-            }
-
-            structured_logger.log_with_context(
-                "INFO", "Operation with sensitive data", **sensitive_context
-            )
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
