+++ tests/test_import_data_filesystem.py
@@ -1,4 +1,4 @@
-"""Final Coverage Push Tests
+"""Import, Data Structures and FileSystem Coverage Tests
 
 Targeted tests to push coverage from 12% toward 20-30%.
 Focus on previously untested but easily testable modules.
@@ -386,314 +386,3 @@ class TestFileSystemCoverage:
                 FileNotFoundError,
             ) as e:
                 pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestPerformanceLoggingCoverage:
-    """Test performance logging for coverage"""
-
-    def test_log_performance_comprehensive(self):
-        """Test log performance comprehensive functionality"""
-        from kumihan_formatter.core.utilities.log_performance import LogPerformance
-
-        perf_logger = LogPerformance()
-
-        # Test performance measurement scenarios
-        operations = [
-            ("parse_document", 0.125),
-            ("render_html", 0.089),
-            ("validate_syntax", 0.034),
-            ("file_io_operation", 0.256),
-            ("template_processing", 0.067),
-        ]
-
-        for operation_name, duration in operations:
-            try:
-                # Log basic performance
-                perf_logger.log_operation(operation_name, duration)
-
-                # Log with additional metrics
-                perf_logger.log_operation_with_metrics(
-                    operation_name,
-                    duration,
-                    cpu_usage=45.2,
-                    memory_usage=128 * 1024 * 1024,  # 128 MB
-                    items_processed=150,
-                )
-
-                # Test performance analysis
-                analysis = perf_logger.analyze_operation(operation_name)
-                assert analysis is not None
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test performance reporting
-        try:
-            report = perf_logger.generate_report()
-            assert report is not None
-
-            # Test performance statistics
-            stats = perf_logger.get_statistics()
-            assert isinstance(stats, dict)
-
-            # Test performance trends
-            trends = perf_logger.get_trends()
-            assert trends is not None
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-    def test_log_size_control_comprehensive(self):
-        """Test log size control comprehensive functionality"""
-        from kumihan_formatter.core.utilities.log_size_control import LogSizeControl
-
-        size_control = LogSizeControl()
-
-        # Test log size management
-        try:
-            # Set size limits
-            size_control.set_max_size(10 * 1024 * 1024)  # 10 MB
-            size_control.set_rotation_count(5)
-
-            # Test size checking
-            current_size = size_control.get_current_size()
-            assert isinstance(current_size, int)
-
-            # Test rotation trigger
-            needs_rotation = size_control.needs_rotation()
-            assert isinstance(needs_rotation, bool)
-
-            # Test cleanup operations
-            size_control.cleanup_old_logs()
-
-            # Test compression
-            size_control.compress_old_logs()
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-        # Test log archiving
-        try:
-            archive_path = size_control.archive_logs()
-            assert archive_path is None or isinstance(archive_path, str)
-
-            # Test archive cleanup
-            size_control.cleanup_archives()
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestTextProcessorDeepCoverage:
-    """Deep coverage for text processor"""
-
-    def test_text_processor_all_methods(self):
-        """Test all text processor methods for coverage"""
-        from kumihan_formatter.core.utilities.text_processor import TextProcessor
-
-        processor = TextProcessor()
-
-        # Comprehensive text test cases
-        text_cases = [
-            # Basic cases
-            "Simple text",
-            "  Text with spaces  ",
-            "Text\nwith\nnewlines",
-            "Text\twith\ttabs",
-            # Special characters
-            "Text & with < special > chars",
-            "Text with 'quotes' and \"double quotes\"",
-            "Text with unicode: 日本語 文字",
-            # HTML content
-            "<p>HTML <strong>content</strong></p>",
-            "<div class='test'>Nested <span>tags</span></div>",
-            # Mixed content
-            "Mixed content:\n<p>HTML</p>\nAnd ;;;kumihan;;; syntax ;;;",
-            # Edge cases
-            "",  # Empty
-            "   ",  # Only spaces
-            "\n\n\n",  # Only newlines
-            "A",  # Single character
-        ]
-
-        # Test all available methods
-        methods_to_test = [
-            "normalize_whitespace",
-            "clean_text",
-            "process_text",
-            "strip_html_tags",
-            "escape_special_chars",
-            "unescape_html",
-            "remove_markdown",
-            "extract_text",
-            "count_words",
-            "count_characters",
-            "split_sentences",
-            "split_paragraphs",
-        ]
-
-        for text in text_cases:
-            for method_name in methods_to_test:
-                if hasattr(processor, method_name):
-                    try:
-                        method = getattr(processor, method_name)
-                        result = method(text)
-
-                        # Basic validation
-                        if method_name.startswith("count_"):
-                            assert isinstance(result, int)
-                        elif method_name.startswith("split_"):
-                            assert isinstance(result, list)
-                        else:
-                            assert isinstance(result, str)
-
-                    except (
-                        AttributeError,
-                        NotImplementedError,
-                        TypeError,
-                        ValueError,
-                        FileNotFoundError,
-                    ) as e:
-                        # Some methods may fail on edge cases
-                        pytest.skip(f"Method or operation not available: {e}")
-
-        # Test text transformation combinations
-        try:
-            sample_text = "  <p>Sample **markdown** text</p>  "
-
-            # Chain operations
-            step1 = processor.normalize_whitespace(sample_text)
-            step2 = processor.strip_html_tags(step1)
-            step3 = processor.clean_text(step2)
-
-            assert isinstance(step3, str)
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestSecurityPatternsCoverage:
-    """Test security patterns for coverage"""
-
-    def test_security_patterns_comprehensive(self):
-        """Test security patterns comprehensive functionality"""
-        from kumihan_formatter.core.utilities.security_patterns import SecurityPatterns
-
-        security = SecurityPatterns()
-
-        # Test input validation scenarios
-        validation_tests = [
-            # Safe inputs
-            ("safe_text", "Hello world"),
-            ("safe_path", "/safe/path/file.txt"),
-            ("safe_html", "<p>Safe content</p>"),
-            # Potentially unsafe inputs
-            ("script_injection", "<script>alert('xss')</script>"),
-            ("path_traversal", "../../../etc/passwd"),
-            ("sql_injection", "'; DROP TABLE users; --"),
-            ("command_injection", "; rm -rf /"),
-        ]
-
-        for test_type, input_data in validation_tests:
-            try:
-                # Test input validation
-                is_safe = security.validate_input(input_data)
-                assert isinstance(is_safe, bool)
-
-                # Test input sanitization
-                sanitized = security.sanitize_input(input_data)
-                assert isinstance(sanitized, str)
-
-                # Test threat detection
-                threat_level = security.detect_threat(input_data)
-                assert threat_level is not None
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test security rules
-        try:
-            # Test rule validation
-            rules = security.get_security_rules()
-            assert isinstance(rules, (list, dict))
-
-            # Test rule application
-            test_input = "<script>test</script>"
-            filtered = security.apply_rules(test_input)
-            assert isinstance(filtered, str)
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-        # Test encoding/escaping
-        try:
-            test_strings = [
-                "Normal text",
-                "<script>alert('test')</script>",
-                "Text with & special < chars >",
-                "Unicode: 日本語",
-            ]
-
-            for test_string in test_strings:
-                # Test HTML escaping
-                escaped = security.escape_html(test_string)
-                assert isinstance(escaped, str)
-
-                # Test URL encoding
-                url_encoded = security.encode_url(test_string)
-                assert isinstance(url_encoded, str)
-
-                # Test safe decoding
-                decoded = security.safe_decode(url_encoded)
-                assert isinstance(decoded, str)
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
