+++ tests/test_parser_integration.py
@@ -1,6 +1,6 @@
-"""Parser and Renderer Integration Tests
+"""Parser and Specific Module Integration Tests
 
-Deep integration tests for parser and renderer systems.
+Deep integration tests for parser systems and specific modules.
 Targets specific uncovered code paths for maximum coverage impact.
 """
 
@@ -453,226 +453,6 @@ Content for section 2."""
         try:
             is_valid = parser.validate_nesting(nested_input)
             assert isinstance(is_valid, bool)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-
-class TestUtilitiesDeepCoverage:
-    """Deep coverage for utility modules"""
-
-    def test_dependency_tracker_functionality(self):
-        """Test dependency tracker functionality"""
-        try:
-            from kumihan_formatter.core.utilities.dependency_tracker import (
-                DependencyTracker,
-            )
-        except ImportError as e:
-            pytest.skip(f"Module not available: {e}")
-            return
-
-        tracker = DependencyTracker()
-
-        # Test dependency registration
-        dependencies = [
-            ("module_a", "module_b"),
-            ("module_b", "module_c"),
-            ("module_a", "module_c"),
-            ("module_d", "module_a"),
-        ]
-
-        for dep_from, dep_to in dependencies:
-            try:
-                tracker.add_dependency(dep_from, dep_to)
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                ImportError,
-            ) as e:
-                pass
-
-        # Test circular dependency detection
-        try:
-            has_circular = tracker.has_circular_dependencies()
-            assert isinstance(has_circular, bool)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-        # Test dependency resolution order
-        try:
-            resolution_order = tracker.get_resolution_order()
-            assert isinstance(resolution_order, (list, tuple))
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-    def test_error_analyzer_functionality(self):
-        """Test error analyzer functionality"""
-        try:
-            from kumihan_formatter.core.utilities.error_analyzer import ErrorAnalyzer
-        except ImportError as e:
-            pytest.skip(f"Module not available: {e}")
-            return
-
-        analyzer = ErrorAnalyzer()
-
-        # Test error analysis
-        test_errors = [
-            Exception("Test error message"),
-            ValueError("Invalid value provided"),
-            TypeError("Type mismatch"),
-            FileNotFoundError("File not found"),
-        ]
-
-        for error in test_errors:
-            try:
-                analysis = analyzer.analyze_error(error)
-                assert analysis is not None
-                assert isinstance(analysis, dict)
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                ImportError,
-            ) as e:
-                pass
-
-        # Test error categorization
-        try:
-            categories = analyzer.categorize_errors(test_errors)
-            assert isinstance(categories, dict)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-        # Test error suggestions
-        try:
-            suggestions = analyzer.get_suggestions("FileNotFoundError")
-            assert isinstance(suggestions, (list, str))
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-    def test_string_similarity_functionality(self):
-        """Test string similarity functionality"""
-        try:
-            from kumihan_formatter.core.utilities.string_similarity import (
-                StringSimilarity,
-            )
-        except ImportError as e:
-            pytest.skip(f"Module not available: {e}")
-            return
-
-        similarity = StringSimilarity()
-
-        # Test similarity calculations
-        test_pairs = [
-            ("hello", "hello"),  # Identical
-            ("hello", "hallo"),  # Similar
-            ("hello", "world"),  # Different
-            ("test", "testing"),  # Substring
-            ("", ""),  # Empty
-            ("a", "b"),  # Single char
-        ]
-
-        for str1, str2 in test_pairs:
-            try:
-                score = similarity.calculate(str1, str2)
-                assert isinstance(score, (int, float))
-                assert 0 <= score <= 1 or 0 <= score <= 100
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                ImportError,
-            ) as e:
-                pass
-
-        # Test fuzzy matching
-        try:
-            matches = similarity.fuzzy_match(
-                "test", ["testing", "text", "rest", "best"]
-            )
-            assert isinstance(matches, list)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-    def test_execution_flow_tracker_functionality(self):
-        """Test execution flow tracker functionality"""
-        try:
-            from kumihan_formatter.core.utilities.execution_flow_tracker import (
-                ExecutionFlowTracker,
-            )
-        except ImportError as e:
-            pytest.skip(f"Module not available: {e}")
-            return
-
-        tracker = ExecutionFlowTracker()
-
-        # Test flow tracking
-        try:
-            tracker.start_tracking()
-
-            # Simulate execution flow
-            tracker.track_function_call("function_a")
-            tracker.track_function_call("function_b")
-            tracker.track_function_return("function_b")
-            tracker.track_function_return("function_a")
-
-            flow_data = tracker.get_flow_data()
-            assert isinstance(flow_data, (dict, list))
-
-            tracker.stop_tracking()
-
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-        # Test performance analysis
-        try:
-            performance_data = tracker.analyze_performance()
-            assert isinstance(performance_data, dict)
         except (
             AttributeError,
             NotImplementedError,
