+++ tests/test_validation_utility_concurrency_logger_error_handling.py
@@ -1,239 +1,15 @@
-"""Error Handling Coverage Tests
+"""Validation, Utility, Concurrency and Logger Error Handling Tests
 
-Focus on error handling, exception management, and edge cases.
+Focus on validation, utilities, concurrency and logger error handling.
 Target: Increase error handling module coverage significantly.
 """
 
 import tempfile
+import threading
 from pathlib import Path
 from unittest.mock import Mock, patch
 
 import pytest
-
-
-class TestParserErrorHandling:
-    """Test parser error handling"""
-
-    def test_parser_malformed_syntax(self):
-        """Test handling of malformed syntax"""
-        from kumihan_formatter.parser import Parser
-
-        parser = Parser()
-
-        # Test malformed Kumihan syntax
-        malformed_cases = [
-            ";;;incomplete;;; missing end",
-            ";;;nested;;; ;;;inner;;; content ;;; ;;;",
-            "((incomplete footnote",
-            "｜incomplete《ruby",
-            ";;;empty;;;  ;;;",
-            "((()))",  # Empty nested parentheses
-            "｜《》",  # Empty ruby
-        ]
-
-        for malformed_text in malformed_cases:
-            try:
-                result = parser.parse(malformed_text)
-                # Should handle gracefully, not crash
-                assert result is not None
-                assert isinstance(result, list)
-            except Exception as e:
-                # Some malformed syntax might raise exceptions
-                # Ensure they are handled appropriately
-                assert isinstance(e, (ValueError, SyntaxError, TypeError))
-
-    def test_parser_edge_cases(self):
-        """Test parser edge cases"""
-        from kumihan_formatter.parser import Parser
-
-        parser = Parser()
-
-        edge_cases = [
-            "",  # Empty string
-            "\n\n\n",  # Only newlines
-            "   \t   ",  # Only whitespace
-            "\x00\x01\x02",  # Control characters
-            "A" * 10000,  # Very long line
-            "\n".join(["Line"] * 1000),  # Many lines
-        ]
-
-        for edge_case in edge_cases:
-            try:
-                result = parser.parse(edge_case)
-                assert result is not None
-                assert isinstance(result, list)
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                # Some edge cases might not be handled
-                pytest.skip(f"Edge case handling not available: {e}")
-
-    def test_parser_unicode_handling(self):
-        """Test Unicode and special character handling"""
-        from kumihan_formatter.parser import Parser
-
-        parser = Parser()
-
-        unicode_cases = [
-            "🎉 Emoji test",
-            "中文测试",  # Chinese
-            "العربية",  # Arabic
-            "Русский",  # Russian
-            "🚀 Mixed ✨ emoji 🌟 text",
-            "Zero-width:\u200b\u200c\u200d",
-            "RTL: \u202e\u202d text",
-        ]
-
-        for unicode_text in unicode_cases:
-            try:
-                result = parser.parse(unicode_text)
-                assert result is not None
-                assert isinstance(result, list)
-            except (UnicodeError, UnicodeDecodeError):
-                # Unicode handling might not be complete
-                pass
-
-
-class TestRendererErrorHandling:
-    """Test renderer error handling"""
-
-    def test_renderer_invalid_nodes(self):
-        """Test handling of invalid nodes"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Test invalid node structures
-        invalid_nodes = [
-            [None],  # None node
-            [Node("", "")],  # Empty node type
-            [Node("invalid_type", "content")],  # Unknown type
-            [Node("p", None)],  # None content
-            [Mock()],  # Mock object instead of Node
-        ]
-
-        for nodes in invalid_nodes:
-            try:
-                result = renderer.render(nodes)
-                # Should handle gracefully
-                assert isinstance(result, str)
-            except (TypeError, AttributeError, ValueError):
-                # Some invalid nodes might raise exceptions
-                pass
-
-    def test_renderer_circular_references(self):
-        """Test handling of circular references"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Create circular reference
-        node1 = Node("div", "")
-        node2 = Node("p", "")
-
-        try:
-            # Create circular structure
-            node1.content = [node2]
-            node2.content = [node1]
-
-            result = renderer.render([node1])
-            # Should detect and handle circular references
-            assert isinstance(result, str)
-
-        except (RecursionError, RuntimeError):
-            # Circular references might cause recursion errors
-            pass
-        except AttributeError:
-            # Content might be read-only
-            pass
-
-    def test_renderer_memory_limits(self):
-        """Test renderer with large structures"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Create very large structure
-        large_nodes = []
-        for i in range(1000):
-            node = Node("p", f"Paragraph {i} with content")
-            large_nodes.append(node)
-
-        try:
-            result = renderer.render(large_nodes)
-            # Should handle large structures
-            assert isinstance(result, str)
-            assert len(result) > 1000
-        except MemoryError:
-            # Large structures might cause memory issues
-            pass
-
-
-class TestConfigErrorHandling:
-    """Test configuration error handling"""
-
-    def test_config_invalid_values(self):
-        """Test handling of invalid configuration values"""
-        from kumihan_formatter.config import ConfigManager
-
-        config = ConfigManager()
-
-        invalid_configs = [
-            {"output_format": None},
-            {"encoding": 123},  # Non-string encoding
-            {"template": []},  # List instead of string
-            {"nested": {"invalid": object()}},  # Non-serializable value
-        ]
-
-        for invalid_config in invalid_configs:
-            try:
-                config.load_config(invalid_config)
-                # Should handle invalid values gracefully
-            except (TypeError, ValueError):
-                # Invalid values might raise exceptions
-                pass
-            except AttributeError:
-                # Method might not exist
-                pass
-
-    def test_config_file_errors(self):
-        """Test configuration file error handling"""
-        from kumihan_formatter.config.config_manager import ConfigManager
-
-        config_manager = ConfigManager()
-
-        # Test non-existent file
-        try:
-            config_manager.load_from_file("non_existent_config.json")
-        except (FileNotFoundError, IOError):
-            # Expected behavior
-            pass
-        except AttributeError:
-            # Method might not exist
-            pass
-
-        # Test malformed JSON
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".json", delete=False) as f:
-            f.write("{invalid json content")
-            malformed_path = f.name
-
-        try:
-            config_manager.load_from_file(malformed_path)
-        except (ValueError, TypeError):  # JSON decode error
-            # Expected behavior
-            pass
-        except AttributeError:
-            # Method might not exist
-            pass
-        finally:
-            Path(malformed_path).unlink(missing_ok=True)
 
 
 class TestValidationErrorHandling:
