+++ tests/test_working_node_utilities.py
@@ -1,4 +1,4 @@
-"""Working Modules Deep Coverage Tests
+"""Working Node Builder and Utilities Deep Coverage Tests
 
 Focus on modules that are confirmed to be working and can be deeply tested.
 Strategy: Exercise existing working code paths extensively.
@@ -10,257 +10,6 @@ from pathlib import Path
 from unittest.mock import Mock, patch
 
 import pytest
-
-
-class TestWorkingParserDeepCoverage:
-    """Deep test working parser functionality"""
-
-    def test_parse_function_extensive(self):
-        """Test parse function with extensive scenarios"""
-        from kumihan_formatter import parse
-
-        # Test extensive parsing scenarios
-        parsing_scenarios = [
-            # Simple cases
-            "Hello world",
-            "",
-            "Single line",
-            # Multi-line cases
-            "Line 1\nLine 2",
-            "Line 1\n\nLine 3",
-            "Multiple\nlines\nof\ncontent",
-            # Heading patterns
-            "# Heading 1",
-            "## Heading 2",
-            "### Heading 3",
-            "#### Heading 4",
-            "##### Heading 5",
-            # Formatting patterns
-            "**Bold text**",
-            "*Italic text*",
-            "***Bold italic***",
-            "`Code text`",
-            # Kumihan patterns (if implemented)
-            ";;;highlight;;; content ;;;",
-            ";;;box;;; content ;;;",
-            "((footnote))",
-            "｜ruby《reading》",
-            # Complex combinations
-            """# Document Title
-
-This is a paragraph with **bold** and *italic* text.
-
-## Section 1
-
-More content here.
-
-;;;highlight;;; Important information ;;;
-
-Another paragraph.""",
-        ]
-
-        for scenario in parsing_scenarios:
-            try:
-                result = parse(scenario)
-                assert result is not None
-
-                # Result should be structured data
-                if hasattr(result, "__len__"):
-                    assert len(result) >= 0
-
-                # If it's a string, should be valid
-                if isinstance(result, str):
-                    assert len(result) >= 0
-
-            except Exception as e:
-                # Log failures but continue testing
-                print(f"Parse failed for: {scenario[:30]}... - {e}")
-
-    def test_render_function_extensive(self):
-        """Test render function with extensive scenarios"""
-        from kumihan_formatter import render
-        from kumihan_formatter.core.ast_nodes.node import Node
-
-        # Test extensive rendering scenarios
-        node_scenarios = [
-            # Simple nodes
-            [Node("p", "Simple paragraph")],
-            [Node("h1", "Heading")],
-            [Node("div", "Container")],
-            # Multiple nodes
-            [
-                Node("h1", "Title"),
-                Node("p", "Paragraph 1"),
-                Node("p", "Paragraph 2"),
-            ],
-            # Nested nodes
-            [
-                Node(
-                    "div",
-                    [
-                        Node("h2", "Section"),
-                        Node("p", "Content"),
-                    ],
-                )
-            ],
-            # Complex structure
-            [
-                Node(
-                    "div",
-                    [
-                        Node("h1", "Main Title"),
-                        Node(
-                            "div",
-                            [
-                                Node("h2", "Subsection"),
-                                Node("p", "Description"),
-                                Node(
-                                    "ul",
-                                    [
-                                        Node("li", "Item 1"),
-                                        Node("li", "Item 2"),
-                                    ],
-                                ),
-                            ],
-                        ),
-                    ],
-                )
-            ],
-            # Different node types
-            [Node("strong", "Bold")],
-            [Node("em", "Italic")],
-            [Node("code", "Code")],
-            [Node("blockquote", "Quote")],
-            [Node("pre", "Preformatted")],
-        ]
-
-        for nodes in node_scenarios:
-            try:
-                result = render(nodes)
-                assert isinstance(result, str)
-                assert len(result) > 0
-
-                # Should contain some representation of content
-                for node in nodes:
-                    if hasattr(node, "content") and isinstance(node.content, str):
-                        if node.content and node.content not in result:
-                            # Content might be transformed, but some representation should exist
-                            pass
-
-            except Exception as e:
-                print(f"Render failed for: {[n.type for n in nodes]} - {e}")
-
-
-class TestWorkingConfigDeepCoverage:
-    """Deep test working config functionality"""
-
-    def test_config_manager_extensive_scenarios(self):
-        """Test config manager with extensive scenarios"""
-        from kumihan_formatter.config.config_manager import ConfigManager
-
-        config_manager = ConfigManager()
-
-        # Test extensive configuration scenarios
-        config_scenarios = [
-            # Basic settings
-            {"encoding": "utf-8"},
-            {"output_format": "html"},
-            {"template": "default"},
-            # Complex settings
-            {
-                "parser": {
-                    "strict_mode": False,
-                    "allow_html": True,
-                    "encoding": "utf-8",
-                },
-                "renderer": {
-                    "template": "default",
-                    "minify": False,
-                    "include_css": True,
-                },
-                "output": {
-                    "format": "html",
-                    "file_extension": ".html",
-                    "encoding": "utf-8",
-                },
-            },
-            # Edge cases
-            {},  # Empty config
-            {"unknown_setting": "value"},  # Unknown setting
-            {"nested": {"deeply": {"nested": {"value": True}}}},  # Deep nesting
-        ]
-
-        for config_data in config_scenarios:
-            try:
-                # Test config loading
-                config_manager.load_config(config_data)
-
-                # Test value retrieval
-                for key in config_data.keys():
-                    value = config_manager.get(key)
-                    # Value should be retrievable (may be transformed)
-
-                # Test config validation if available
-                if hasattr(config_manager, "validate"):
-                    is_valid = config_manager.validate()
-                    assert isinstance(is_valid, bool)
-
-                # Test config export
-                if hasattr(config_manager, "export"):
-                    exported = config_manager.export()
-                    assert isinstance(exported, dict)
-
-            except Exception as e:
-                print(f"Config test failed for: {config_data} - {e}")
-
-    def test_base_config_extensive(self):
-        """Test base config with extensive scenarios"""
-        from kumihan_formatter.config.base_config import BaseConfig
-
-        base_config = BaseConfig()
-
-        # Test various configuration operations
-        config_operations = [
-            ("set", "key1", "value1"),
-            ("set", "nested.key", "nested_value"),
-            ("set", "number", 42),
-            ("set", "boolean", True),
-            ("set", "list", [1, 2, 3]),
-            ("set", "dict", {"inner": "value"}),
-        ]
-
-        for operation, *args in config_operations:
-            try:
-                if operation == "set" and hasattr(base_config, "set"):
-                    base_config.set(*args)
-
-                    # Test retrieval
-                    if hasattr(base_config, "get"):
-                        retrieved_value = base_config.get(args[0])
-                        # Should be able to retrieve the value
-
-            except Exception as e:
-                print(f"Base config operation failed: {operation} {args} - {e}")
-
-        # Test config merging
-        try:
-            if hasattr(base_config, "merge"):
-                merge_data = {"new_key": "new_value", "override": "override_value"}
-                base_config.merge(merge_data)
-
-        except Exception:
-            pass
-
-        # Test config sections
-        try:
-            if hasattr(base_config, "create_section"):
-                base_config.create_section("test_section")
-                base_config.set_section_value(
-                    "test_section", "section_key", "section_value"
-                )
-
-        except Exception:
-            pass
 
 
 class TestWorkingNodeBuilderDeepCoverage:
